agent-doc-cli/pyproject.toml
[project]
name = "agent-doc-cli"
version = "0.1.0"
description = "Document management agent built with Google ADK"
dependencies = [
    "google-adk>=1.15.0,<2.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.9",
]
requires-python = ">=3.10,<3.14"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
